requirements.txt
python-dotenv
Flask
langchain
langchain-community
langchain-core
langchain-experimental
pyngrok